Prepare player publisher API
